One map landing page automation scripts
For Landing page added below scenario
# Landing Page Title Validation:
Verifies the expected title when the landing page loads.
# Landing Page Heading Validation:
Ensures the expected heading is present upon the landing page loading.
# Landing Page Logo Verification:
Checks for the presence of the expected logo upon landing page loading.
#Top Center Menu Validation:
Validates the presence and correctness of all menus located at the top center of the landing page.
#Footer Right Corner Links Validation:
Confirms the presence and accuracy of all links situated at the footer's right corner.
#View Changes between 2D and 3D Verification:
Verifies the smooth transition between 2D and 3D views and vice versa.
#"Help Us Improve" Functionality Verification:
Checks the functionality enabling users to provide ratings and comments for improvement.